auto_inchis

+++ dataloader.py
@@ -47,7 +47,7 @@ class MoleculeDataset(Dataset):
 
         ### grab inchi
         inchi_idx = i + (200000*self.shard_id)
-        inchi_data = torch.tensor(self.encoded_inchis[inchi_idx]).int()
+        inchi_data = torch.tensor(self.encoded_inchis[inchi_idx]).long()
         encoded_inchi = inchi_data[:-1]
         inchi_length = inchi_data[-1]
         return img, encoded_inchi, inchi_length
